chore: refactor types out into separate named files
